requirements.txt
amqp==5.2.0
asgiref==3.8.1
astroid==3.1.0
billiard==4.2.0
black==24.4.0
bsdiff4==1.2.4
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
Cython==3.0.10
dill==0.3.8
Django==4.2.11
djangorestframework==3.15.1
docker==7.0.0
drf-yasg==1.21.7
honeybadger==0.19.1
idna==3.7
inflection==0.5.1
isort==5.13.2
jsonpickle==3.0.4
kombu==5.3.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
psycopg2-binary==2.9.9
pycparser==2.22
pycrunch-trace==0.2
pylint==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
